pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clv-retention-engine"
version = "0.1.0"
description = "Production-grade CLV retention decision engine"
requires-python = ">=3.11"
dependencies = [
  "duckdb>=1.0.0",
  "pandas>=2.1.0",
  "numpy>=1.26.0",
  "scikit-learn>=1.4.0",
  "xgboost>=2.0.0",
  "fastapi>=0.110.0",
  "uvicorn>=0.29.0",
  "streamlit>=1.32.0",
  "pyyaml>=6.0.1",
  "joblib>=1.3.2",
  "matplotlib>=3.8.0",
  "httpx>=0.27.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
